Update test scripts
