+++ packages/salesforcedx-slds-linter/package.json
@@ -36,7 +36,7 @@
     "clean": "shx rm -rf node_modules",
     "postinstall": "node ./node_modules/vscode/bin/install",
     "test": "node ./node_modules/vscode/bin/test",
-    "test:unit": "node ./node_modules/vscode/bin/test"
+    "test:vscode-integration": "node ./node_modules/vscode/bin/test"
   },
   "nyc": {
     "reporter": ["text-summary", "lcov"]
