+++ packages/salesforcedx-utils-vscode/test/unit/cli/forceConfigGet.test.ts
@@ -7,62 +7,54 @@
 
 import { expect } from 'chai';
 import * as sinon from 'sinon';
-import { ForceOrgDisplay, OrgInfo, SfdxCommandBuilder } from '../../src/cli';
+import { ForceConfigGet, SfdxCommandBuilder } from '../../../src/cli';
 import childProcess = require('child_process');
 
-describe('force:org:display', () => {
+describe('force:config:get', () => {
   const mockSpawn = require('mock-spawn');
-  let command: ForceOrgDisplay;
+  let command: ForceConfigGet;
   let origSpawn: any;
   let mySpawn: any;
   let cmdWithArgSpy: sinon.SinonSpy;
-  let cmdWithJsonSpy: sinon.SinonSpy;
+  let cmdJsonSpy: sinon.SinonSpy;
   let cmdBuildSpy: sinon.SinonSpy;
 
   beforeEach(() => {
-    command = new ForceOrgDisplay();
+    command = new ForceConfigGet();
     origSpawn = childProcess.spawn;
     mySpawn = mockSpawn();
     childProcess.spawn = mySpawn;
     cmdWithArgSpy = sinon.spy(SfdxCommandBuilder.prototype, 'withArg');
-    cmdWithJsonSpy = sinon.spy(SfdxCommandBuilder.prototype, 'withJson');
+    cmdJsonSpy = sinon.spy(SfdxCommandBuilder.prototype, 'withJson');
     cmdBuildSpy = sinon.spy(SfdxCommandBuilder.prototype, 'build');
   });
 
   afterEach(() => {
     childProcess.spawn = origSpawn;
     cmdWithArgSpy.restore();
-    cmdWithJsonSpy.restore();
+    cmdJsonSpy.restore();
     cmdBuildSpy.restore();
   });
 
-  it('Should return org info successfully', async () => {
-    const orgInfo: OrgInfo = {
-      username: 'name',
-      devHubId: 'devHubId',
-      id: 'id',
-      createdBy: 'someone',
-      createdDate: new Date().toDateString(),
-      expirationDate: new Date().toDateString(),
-      status: 'active',
-      edition: 'Enterprise',
-      orgName: 'My org',
-      accessToken: '123',
-      instanceUrl: 'https://wwww.salesforce.com',
-      clientId: 'foo'
-    };
+  it('Should return config successfully', async () => {
+    const response = new Array<any>();
+    response.push({ key: 'key1', value: 'val1' });
+    response.push({ key: 'key2', value: 'val2' });
     mySpawn.setDefault(
-      mySpawn.simple(0, `{ "status": 0, "result": ${JSON.stringify(orgInfo)}}`)
+      mySpawn.simple(0, `{ "status": 0, "result": ${JSON.stringify(response)}}`)
     );
 
-    const cmdOutput: OrgInfo = await command.getOrgInfo('foo');
+    const cmdOutput = await command.getConfig('foo', 'key1', 'key2');
 
-    expect(cmdOutput).to.deep.equal(orgInfo);
-    expect(cmdWithArgSpy.calledOnce).to.equal(true);
+    expect(cmdOutput.get('key1')).to.equal('val1');
+    expect(cmdOutput.get('key2')).to.equal('val2');
+    expect(cmdWithArgSpy.callCount).to.equal(3);
     expect(cmdWithArgSpy.getCall(0).args).to.have.same.members([
-      'force:org:display'
+      'force:config:get'
     ]);
-    expect(cmdWithJsonSpy.calledOnce).to.equal(true);
+    expect(cmdWithArgSpy.getCall(1).args).to.have.same.members(['key1']);
+    expect(cmdWithArgSpy.getCall(2).args).to.have.same.members(['key2']);
+    expect(cmdJsonSpy.calledOnce).to.equal(true);
     expect(cmdBuildSpy.calledOnce).to.equal(true);
   });
 
@@ -70,7 +62,7 @@ describe('force:org:display', () => {
     mySpawn.setDefault(mySpawn.simple(1, '', 'There was an error'));
 
     try {
-      await command.getOrgInfo('foo');
+      await command.getConfig('foo');
     } catch (error) {
       expect(error).to.equal('There was an error');
     }
@@ -80,9 +72,10 @@ describe('force:org:display', () => {
     mySpawn.setDefault(mySpawn.simple(0, '{ not valid JSON'));
 
     try {
-      await command.getOrgInfo('foo');
+      await command.getConfig('foo');
+      expect(false).to.equal('should not reach this point!');
     } catch (error) {
-      expect(error).to.equal('{ not valid JSON');
+      // good
     }
   });
 });
