+++ packages/salesforcedx-utils-vscode/test/unit/requestService/requestService.test.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect } from 'chai';
-import { RequestService } from '../../src/requestService';
+import { RequestService } from '../../../src/requestService';
 
 describe('RequestService', () => {
   describe('ENV Variables', () => {
